Add map extension
